backend/requirements.txt
Flask
requests
python-dotenv
WeasyPrint
flask-cors
beautifulsoup4